aapp/requirements.txt
streamlit
openai
plotly
pandas
python-dotenv
Pillow
numpy
